feat: create reply thread
There is now an additional question to add more posts as a reply thread.
Create a temporary folder to copy media files and at the end, rename the
folder to the scheduled time. This is to simplify the code logic but I
want to improve and simplify this part later.

Uploading is not yet updated to support reply threads yet.